refactor(ui actions): create base, electron and web versions of actions/ui.ts file
