+++ app/actions/ui.web.ts
@@ -1,0 +1,8 @@
+import { UI_SIGNOUT } from '../reducers/ui'
+export * from './ui.base'
+
+export const signout = () => {
+  return {
+    type: UI_SIGNOUT
+  }
+}
