+++ app/components/compare/CompareSidebar.tsx
@@ -8,7 +8,7 @@ import Store, { ToastType } from '../../models/store'
 
 import { connectComponentToProps } from '../../utils/connectComponentToProps'
 
-import { openToast } from '../../actions/ui'
+import { openToast } from '../../actions/ui.TARGET_PLATFORM'
 
 import TextInput from '../form/TextInput'
 import ButtonInput from '../form/ButtonInput'
