+++ app/components/modals/AddDataset.tsx
@@ -7,7 +7,7 @@ import { validateDatasetReference } from '../../utils/formValidation'
 import { connectComponentToProps } from '../../utils/connectComponentToProps'
 
 import { addDatasetAndFetch } from '../../actions/api'
-import { dismissModal } from '../../actions/ui'
+import { dismissModal } from '../../actions/ui.TARGET_PLATFORM'
 
 import Modal from './Modal'
 import ExternalLink from '../ExternalLink.TARGET_PLATFORM'
