+++ app/components/workbench/NoDatasetsMainContent.tsx
@@ -5,7 +5,7 @@ import { Link } from 'react-router-dom'
 
 import { connectComponentToProps } from '../../utils/connectComponentToProps'
 
-import { setModal } from '../../actions/ui'
+import { setModal } from '../../actions/ui.TARGET_PLATFORM'
 
 import { Modal, ModalType } from '../../models/modals'
 import WelcomeTemplate from '../onboard/WelcomeTemplate.TARGET_PLATFORM'
